requirements.txt
Django
pillow
django-extensions
django-robots
django-debug-toolbar
django-taggit
django-summernote
django-simple-captcha
django-multi-captcha-admin
